# src/main/java/io/github/diegorscs/dto/GenderResponse.java
package io.github.diegorscs.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class GenderResponse {
    private String name;
    private String description;
}